requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydub==0.25.1
requests==2.32.3
rsa==4.9.1
urllib3==2.4.0
django-allauth
PyJWT
cryptography
